Added a player health manager
Created a player health manager that will listen to all events for added and removing health from the player and will also kill the player through an event or directly. I am waiting on PR from kyle to tie this to the enemy class he is making and I am waiting on the PR to be accepted from Jackie to tie this to the healing item object.