Base enemy, enemy state, and meleeEnemy class all created
Base enemy has been created, pulled pieces of code from the first base class written in the previous version of the project. The melee enemy class was created with all the necesary override methods and they all have no implemented exceptions in them. I also used the enemy state  enum that DAC made in the previous iteration. I got the enemy states to function correctly (as far as I know) but this is to be tested as I further develop this enemy.