Added enemy bullet/projectile
Created an enemy bullet that fires on an interval when in range of the player. The bullet properly communicates with playerhealthmanager and knockback is also applied from it. This enemy bullet does not collide with enemies. only the player. Still need to add movement to the ranged enemy and look into the enemy states for the ranged enemy