Working with the pause menu
Attempting to tie in the pause menu in all scenes using prefabs. Possibly needs to be reworked. Currently the buttons in the pause menu do not work in any scene except for the main menu scene. Thinking of making 2 seperate options menus. One for the main menu scene and the other for when you are in a game (going off memory this seems like how it is done in most games)