Holy shit it finally is working normally
I FINALLY got the pause menu to function properly. what a disaster to get that going. The game properly pauses when pressing escape when you are in the game. If you are in the main menu scene, the pause function will not work and this means that you are in the main menu so to access options, you press the options button. So happy rn omg