weapon pickup is functional
the actual weapon prefabs are not designed yet but the pickup is functional.

The pickup uses a scriptable object called weapon pickup that contains the weapon sprite, weaponPrefab, and the weaponname.

There is a script called WEaponPickupBehavior that handles the pickup logic.

This pickup script is on the weaponPickup prefab and has a field to hold a weaponSO for a specific weapon. With this design there will be a weaponSO for each weapon as well as a WeaponPickupPrefab for each weapon. Once these are created, they can easily be instantiated in a scene.

The pickup functionality works so when you pick up the weapon, the pickup gets destroyed. Since there is no actual weapon prefab yet, the weapon does not get instantiated onto the player YET. This will be refined a little bit then I will work on the weapon prefabs!