LOTS of refactoring things and melee weapon works
I had to do quite a bit of refactoring to get the new weapon system to attack properly

I have only implemented the melee weapon and this weapon has a trigger method that deals damage to any enemy object. Now, any object that needs to damage an enemy will call the public TakeDamage method in baseEnemy with passed in dmgAmount.

I also realized that the base weapon class and the derived meleeWeapon class from it were completely redundant and not being used so I commented them out for now. The weaponBehavior script handles weapon behaviors such as attacking, attack cooldown, and dealing damage. And this references the weapon SO which has all the specific weapon data. This means there should be no need for the basee weapon class that I loosely created.

Everything runs as intended for what we currently have